auto-improve: revise per review comments
Refs damien-robotsix/robotsix-cai#808
